requirements.txt
google-generativeai>=0.3.0
click>=8.1.0
colorama>=0.4.6
python-dotenv>=1.0.0
pytest>=7.4.0
ast-tools
typing_extensions>=4.7.0
groq>=0.4.0
transformers==4.40.0
sentence-transformers==2.2.2
faiss-cpu
tree-sitter==0.20.4
tree-sitter-languages==1.10.2
torch 
numpy
fastapi
uvicorn